requirements.txt
# Pinned dependencies for reproducibility

# Core ML & Transformers
torch==2.3.1
torchvision==0.18.1
transformers==4.43.3
datasets==2.19.1
scikit-learn==1.4.2

# Serving
fastapi==0.110.1
uvicorn[standard]==0.29.0

# MLOps & Utilities
mlflow==2.14.1
pandas==2.2.2
pydantic==2.8.2
PyYAML==6.0.1
optuna==3.6.1

# Optional for plotting/analysis
dataframe-image==0.2.3
seaborn==0.13.2
matplotlib==3.8.4